libs/text-to-speech/pyproject.toml
[project]
name = "bb-lib-text-to-speech"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9, <3.12"
dependencies = [
    "coqui-tts==0.25.0",
    "torch==2.1.2",
    "torchaudio==2.1.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["bb.lib.text_to_speech"] 